Rename opentracing.mocktracer to opentracing.mock.
This would make it more similar to how it's named
in the Java API, and match the own Python's 'mock'
module.
