+++ opentracing/mock/binary_propagator.py
