+++ opentracing/mock/span.py
